requirements.txt
# requirements.txt
fastapi==0.109.2
uvicorn==0.27.0
requests==2.31.0
python-dateutil==2.8.2
pandas==2.0.3
pybase64==1.3.1
python-dotenv==1.0.1
httpx==0.27.0
markdown==3.6
Pillow==10.2.0